tests: Create dedicated image layout test file
