+++ tests/CMakeLists.txt
@@ -127,6 +127,8 @@ target_sources(vk_layer_validation_tests PRIVATE
     unit/host_image_copy_positive.cpp
     unit/image_drm.cpp
     unit/image_drm_positive.cpp
+    unit/image_layout.cpp
+    unit/image_layout_positive.cpp
     unit/image.cpp
     unit/image_positive.cpp
     unit/imageless_framebuffer.cpp
