+++ tests/unit/push_descriptor.cpp
@@ -244,89 +244,6 @@ TEST_F(NegativePushDescriptor, TemplateDestroyDescriptorSetLayout) {
     vk::DestroyPipelineLayout(device(), pipeline_layout, nullptr);
 }
 
-TEST_F(NegativePushDescriptor, ImageLayout) {
-    TEST_DESCRIPTION("Use a push descriptor with a mismatched image layout.");
-
-    AddRequiredExtensions(VK_KHR_PUSH_DESCRIPTOR_EXTENSION_NAME);
-    RETURN_IF_SKIP(Init());
-    InitRenderTarget();
-
-    VkDescriptorSetLayoutBinding dsl_binding = {};
-    dsl_binding.binding = 0;
-    dsl_binding.descriptorType = VK_DESCRIPTOR_TYPE_COMBINED_IMAGE_SAMPLER;
-    dsl_binding.descriptorCount = 1;
-    dsl_binding.stageFlags = VK_SHADER_STAGE_FRAGMENT_BIT;
-    dsl_binding.pImmutableSamplers = NULL;
-
-    const vkt::DescriptorSetLayout ds_layout(*m_device, {dsl_binding}, VK_DESCRIPTOR_SET_LAYOUT_CREATE_PUSH_DESCRIPTOR_BIT_KHR);
-    auto pipeline_layout = vkt::PipelineLayout(*m_device, {&ds_layout});
-
-    char const *fsSource = R"glsl(
-        #version 450
-        layout(set=0, binding=0) uniform sampler2D tex;
-        layout(location=0) out vec4 color;
-        void main(){
-           color = textureLod(tex, vec2(0.5, 0.5), 0.0);
-        }
-    )glsl";
-    VkShaderObj fs(this, fsSource, VK_SHADER_STAGE_FRAGMENT_BIT);
-
-    CreatePipelineHelper pipe(*this);
-    pipe.shader_stages_[1] = fs.GetStageCreateInfo();
-    pipe.gp_ci_.layout = pipeline_layout.handle();
-    pipe.CreateGraphicsPipeline();
-
-    vkt::Sampler sampler(*m_device, SafeSaneSamplerCreateInfo());
-
-    vkt::Image image(*m_device, 32, 32, 1, VK_FORMAT_B8G8R8A8_UNORM,
-                     VK_IMAGE_USAGE_SAMPLED_BIT | VK_IMAGE_USAGE_COLOR_ATTACHMENT_BIT);
-    vkt::ImageView image_view = image.CreateView();
-    image.SetLayout(VK_IMAGE_ASPECT_COLOR_BIT, VK_IMAGE_LAYOUT_GENERAL);
-
-    VkDescriptorImageInfo img_info = {};
-    img_info.sampler = sampler.handle();
-    img_info.imageView = image_view;
-    img_info.imageLayout = VK_IMAGE_LAYOUT_SHADER_READ_ONLY_OPTIMAL;
-    VkWriteDescriptorSet descriptor_write = vku::InitStructHelper();
-    descriptor_write.dstSet = 0;
-    descriptor_write.descriptorCount = 1;
-    descriptor_write.descriptorType = VK_DESCRIPTOR_TYPE_COMBINED_IMAGE_SAMPLER;
-    descriptor_write.pImageInfo = &img_info;
-    descriptor_write.dstArrayElement = 0;
-    descriptor_write.dstBinding = 0;
-
-    for (uint32_t i = 0; i < 2; i++) {
-        m_command_buffer.Begin();
-        if (i == 1) {
-            // Test path where image layout in command buffer is known at draw time
-            image.ImageMemoryBarrier(m_command_buffer, VK_IMAGE_ASPECT_COLOR_BIT, VK_ACCESS_SHADER_WRITE_BIT,
-                                     VK_ACCESS_SHADER_WRITE_BIT, VK_IMAGE_LAYOUT_COLOR_ATTACHMENT_OPTIMAL,
-                                     VK_PIPELINE_STAGE_FRAGMENT_SHADER_BIT, VK_PIPELINE_STAGE_FRAGMENT_SHADER_BIT);
-        }
-        m_command_buffer.BeginRenderPass(m_renderPassBeginInfo);
-        vk::CmdBindPipeline(m_command_buffer.handle(), VK_PIPELINE_BIND_POINT_GRAPHICS, pipe.Handle());
-        vk::CmdPushDescriptorSetKHR(m_command_buffer.handle(), VK_PIPELINE_BIND_POINT_GRAPHICS, pipeline_layout.handle(), 0, 1,
-                                    &descriptor_write);
-
-        if (i == 1) {
-            // Test path where image layout in command buffer is known at draw time
-            m_errorMonitor->SetDesiredError("VUID-vkCmdDraw-None-08114");
-            m_errorMonitor->SetDesiredError("VUID-VkDescriptorImageInfo-imageLayout-00344");
-            vk::CmdDraw(m_command_buffer.handle(), 1, 1, 0, 0);
-            m_errorMonitor->VerifyFound();
-            break;
-        }
-        m_errorMonitor->SetDesiredError("VUID-vkCmdDraw-None-09600");
-        vk::CmdDraw(m_command_buffer.handle(), 1, 1, 0, 0);
-        m_command_buffer.EndRenderPass();
-        m_command_buffer.End();
-
-        m_default_queue->Submit(m_command_buffer);
-        m_default_queue->Wait();
-        m_errorMonitor->VerifyFound();
-    }
-}
-
 TEST_F(NegativePushDescriptor, SetLayoutWithoutExtension) {
     TEST_DESCRIPTION("Create a push descriptor set layout without loading the needed extension.");
     RETURN_IF_SKIP(Init());
